+++ sdk/resourcemanager/policyinsights/armpolicyinsights/CHANGELOG.md
@@ -1,5 +1,38 @@
 # Release History
 
+## 0.7.0 (2022-10-26)
+### Features Added
+
+- New const `ComponentPolicyStatesResourceLatest`
+- New type alias `ComponentPolicyStatesResource`
+- New function `*ComponentPolicyStatesClient.ListQueryResultsForSubscriptionLevelPolicyAssignment(context.Context, string, string, ComponentPolicyStatesResource, *ComponentPolicyStatesClientListQueryResultsForSubscriptionLevelPolicyAssignmentOptions) (ComponentPolicyStatesClientListQueryResultsForSubscriptionLevelPolicyAssignmentResponse, error)`
+- New function `*ComponentPolicyStatesClient.ListQueryResultsForResource(context.Context, string, ComponentPolicyStatesResource, *ComponentPolicyStatesClientListQueryResultsForResourceOptions) (ComponentPolicyStatesClientListQueryResultsForResourceResponse, error)`
+- New function `PossibleComponentPolicyStatesResourceValues() []ComponentPolicyStatesResource`
+- New function `*ComponentPolicyStatesClient.ListQueryResultsForResourceGroupLevelPolicyAssignment(context.Context, string, string, string, ComponentPolicyStatesResource, *ComponentPolicyStatesClientListQueryResultsForResourceGroupLevelPolicyAssignmentOptions) (ComponentPolicyStatesClientListQueryResultsForResourceGroupLevelPolicyAssignmentResponse, error)`
+- New function `NewComponentPolicyStatesClient(azcore.TokenCredential, *arm.ClientOptions) (*ComponentPolicyStatesClient, error)`
+- New function `*ComponentPolicyStatesClient.ListQueryResultsForResourceGroup(context.Context, string, string, ComponentPolicyStatesResource, *ComponentPolicyStatesClientListQueryResultsForResourceGroupOptions) (ComponentPolicyStatesClientListQueryResultsForResourceGroupResponse, error)`
+- New function `*ComponentPolicyStatesClient.ListQueryResultsForSubscription(context.Context, string, ComponentPolicyStatesResource, *ComponentPolicyStatesClientListQueryResultsForSubscriptionOptions) (ComponentPolicyStatesClientListQueryResultsForSubscriptionResponse, error)`
+- New function `*ComponentPolicyStatesClient.ListQueryResultsForPolicyDefinition(context.Context, string, string, ComponentPolicyStatesResource, *ComponentPolicyStatesClientListQueryResultsForPolicyDefinitionOptions) (ComponentPolicyStatesClientListQueryResultsForPolicyDefinitionResponse, error)`
+- New struct `ComponentExpressionEvaluationDetails`
+- New struct `ComponentPolicyEvaluationDetails`
+- New struct `ComponentPolicyState`
+- New struct `ComponentPolicyStatesClient`
+- New struct `ComponentPolicyStatesClientListQueryResultsForPolicyDefinitionOptions`
+- New struct `ComponentPolicyStatesClientListQueryResultsForPolicyDefinitionResponse`
+- New struct `ComponentPolicyStatesClientListQueryResultsForResourceGroupLevelPolicyAssignmentOptions`
+- New struct `ComponentPolicyStatesClientListQueryResultsForResourceGroupLevelPolicyAssignmentResponse`
+- New struct `ComponentPolicyStatesClientListQueryResultsForResourceGroupOptions`
+- New struct `ComponentPolicyStatesClientListQueryResultsForResourceGroupResponse`
+- New struct `ComponentPolicyStatesClientListQueryResultsForResourceOptions`
+- New struct `ComponentPolicyStatesClientListQueryResultsForResourceResponse`
+- New struct `ComponentPolicyStatesClientListQueryResultsForSubscriptionLevelPolicyAssignmentOptions`
+- New struct `ComponentPolicyStatesClientListQueryResultsForSubscriptionLevelPolicyAssignmentResponse`
+- New struct `ComponentPolicyStatesClientListQueryResultsForSubscriptionOptions`
+- New struct `ComponentPolicyStatesClientListQueryResultsForSubscriptionResponse`
+- New struct `ComponentPolicyStatesQueryResults`
+- New field `IsDataAction` in struct `Operation`
+
+
 ## 0.6.0 (2022-10-07)
 ### Features Added